mv serializable to output_adapter
This keeps mod.rs nice and clean
